% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Perform conversions to kg/h for common units}
\usage{
convert_units(emission_rates, units)
}
\arguments{
\item{emission_rates}{vector of emission rates}

\item{units}{a string describing the units of the emission rate}
}
\value{
A vector of the emission rates converted to kg/h as required for calc_inventory
}
\description{
Perform conversions to kg/h for common units
}
