% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Cleaned MERC/BCOGRIS data}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1701 rows and 19 columns.
}
\usage{
dat
}
\description{
A dataset containing the cleaned dataset from BC. The variables are as follows:
}
\details{
\itemize{
\item anonSiteID site ID
\item anonSourceID component ID
\item anonEquipmentID type of equipment
\item daysSinceInitial number of days since initial survey on that component
\item altitude_m altitude in m
\item windSpeed_ms wind speed in m/s
\item emissionRate emission rate in TODO finish
}
}
\keyword{datasets}
