% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_one_inventory.R
\name{do_one_inventory}
\alias{do_one_inventory}
\title{Calculate one inventory}
\usage{
do_one_inventory(data, est_type)
}
\arguments{
\item{data}{dataframe which has been pre-processed so it has phi, r calculated}

\item{est_type}{the estimate type}
}
\description{
Calculate one inventory
}
