% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_y.R
\name{draw_y}
\alias{draw_y}
\title{simulate a true y value given a msmt}
\usage{
draw_y(Ymeas)
}
\arguments{
\item{Ymeas}{a vector of measured Y values}
}
\value{
A vector of Y values drawn from the distribution. Y values that were equal to zero stay as zero, negative values return NaN.
}
\description{
simulate a true y value given a msmt
}
