% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_phi.R
\name{estimate_phi}
\alias{estimate_phi}
\title{Impute probabilites and estimate phi}
\usage{
estimate_phi(p, num_passes, truncate = 0)
}
\arguments{
\item{p}{vector of known probabilities}

\item{num_passes}{total number of passes}

\item{truncate}{if estimated phi is less than truncate it is set to truncate}
}
\value{
the estimate of phi
}
\description{
Impute probabilites and estimate phi
}
