% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht_variance.R
\name{ht_variance_est}
\alias{ht_variance_est}
\title{Calculate HT style variance estimator}
\usage{
ht_variance_est(p, Ytrue, phi, n_pass)
}
\arguments{
\item{p}{vector of pass-level probabilities}

\item{Ytrue}{pass-level response values}

\item{phi}{estimate or true value of phi}

\item{n_pass}{the total number of passes}
}
\description{
Calculate HT style variance estimator
}
