% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_check.R
\name{mc_check}
\alias{mc_check}
\title{Make plots to assess the number of MC iterations}
\usage{
mc_check(inventory, ...)
}
\arguments{
\item{inventory}{an inventory obtained by running calc_inventory which had measurement error}

\item{...}{additional arguments passed to par}
}
\description{
Make plots to assess the number of MC iterations
}
