% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod_func.R
\name{pod_func}
\alias{pod_func}
\title{Calculate the probability of detection using the function in Conrad et al 2023}
\usage{
pod_func(Ytrue, h, u, truncate = 0.01)
}
\arguments{
\item{Ytrue}{true value of emission rate in kg/h}

\item{h}{height in m}

\item{u}{mind speed in m/s}

\item{truncate}{truncate probabilities less than this value}
}
\description{
Calculate the probability of detection using the function in Conrad et al 2023
}
