% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_inventory.R
\name{calc_inventory}
\alias{calc_inventory}
\title{Estimate methane emissions using a multi-stage approach}
\usage{
calc_inventory(
  data_frame,
  est_type,
  with_me = FALSE,
  n_sim = NA,
  truncate_pod = 0.02,
  bias_corr = 0.918,
  consider_stageII = TRUE
)
}
\arguments{
\item{data_frame}{a data_frame that contains all information on the relevant dataset}

\item{est_type}{string indicating which estimator should be used to aggregate pass data, either "ipw" or "hajek"}

\item{with_me}{logical indicating if measurement error should be incorporated or not}

\item{n_sim}{the number of times we want to simulate y values. Must be specified if with_me = T}

\item{truncate_pod}{value at which to truncate probabilities of detection for small probabilities. Default is 0.02}

\item{bias_corr}{constant to multiply all Y values by to correct for bias. Set to 1 if no bias correction is necessary. Ignored if with_me = TRUE.}

\item{consider_stageII}{Logical; should stage II uncertainty be considered? If TRUE, a simple random sample with population size derived from column Di in data_frame is used. Otherwise, ignored.}
}
\value{
A list:
stratum_totals - point estimates of stratum totals, in kt/y
pop_total - point estimate of population total in kt/y
stratum_var a data.frame with columns var_me (variance due to measurement error), var_de (variance due to the design), and var_total (total variance, var_me + var_de)
pop_var same as stratum_var but for the population total
}
\description{
Estimate methane emissions using a multi-stage approach
}
