% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inventory_info.R
\name{get_inventory_info}
\alias{get_inventory_info}
\title{Take a calc_inventory object and produce CIs for the pop total, and desired individual strata, and variance breakdowns}
\usage{
get_inventory_info(result, strata = T, strata_names = "ALL", conf_level = 0.95)
}
\arguments{
\item{result}{the result of calc_inventory}

\item{strata}{logical; T if CIs for strata are desired}

\item{strata_names}{vector of string names of strata to produce CIs for; required if strata = T. Can set to "ALL" if CIs for all strata are desired}

\item{conf_level}{confidence level}
}
\description{
Take a calc_inventory object and produce CIs for the pop total, and desired individual strata, and variance breakdowns
}
