% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare raw data for analysis}
\usage{
prep_data(
  raw_data,
  component_name,
  facility_name,
  stratum_name,
  day_name,
  nh_name,
  Nh_name,
  Y_name,
  u_name,
  h_name,
  num_wells_name,
  add_day_pop = F,
  day_pop_value = 365
)
}
\arguments{
\item{raw_data}{data.frame containing the raw data (for example, use the dat object included in this package)}

\item{component_name}{name of the column with component identifiers}

\item{facility_name}{name of the column with facility identifiers}

\item{stratum_name}{name of the column with stratum names}

\item{day_name}{name of the column identifying the days}

\item{nh_name}{name of the column with stratum sample sizes}

\item{Nh_name}{name of the column with stratum population sizes}

\item{Y_name}{name of the column with emission rate measurements, in kg/h}

\item{u_name}{name of the column with wind speed measurements, in m/s}

\item{h_name}{name of the column with height measurements, in m}

\item{num_wells_name}{name of the column with the number of wells at the facility}

\item{add_day_pop}{Logical; default is FALSE, and no new columns are added}

\item{day_pop_value}{The number of days to use as the stage II population size when adding a column. Default is 365}
}
\value{
dataframe compatible with calc_inventory; adds column num_passes, and optionally, additional column D if add_day_pop = TRUE
}
\description{
Prepare raw data for analysis
}
